Modification script de rechargement des entites

+++ bin/refresh_entities.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/sh
-#Fonction de logging
+#Rechargement du .bash_profile
+. ~/.bash_profile
+
+#Fonction de loggin
 log_message()
 {
 	vSEV=$1
@@ -7,6 +10,13 @@ log_message()
 	echo `date +'%Y-%m-%d %H:%M:%S'`" [${vSEV}]: ${vMSG}" | tee -a ${LOG_FILE}
 }
 
+if [ -z $WORK_DIR ] ; then
+	log_message ERROR "Variable d'environnement $WORK_DIR non initialisee."
+	log_message ERROR "Veuillez initaliser cette variable dans le fichier ~/.bash_profile"
+	exit 1
+fi
+
+log_message NORMAL "Repertoire de travail : ${WORK_DIR}"
 log_message NORMAL "Suppression du repertoire des Entites"
 rm -rf $WORK_DIR/src/AppBundle/Entity
 cd $WORK_DIR
